Finalized bug fixes
In the browser I was using local storage didn't persist between pages
so there were a lot of "work arounds" I had to incorporate. Those didn't
work on the github pages deployment (because of the above stated).
Refactored local storage variables and parts of the script to represent
consistent data and data manipulation.